grpc_service_debug/environment/requirements.txt
grpcio==1.60.0
protobuf==4.25.0
